pi_deployment/AI_Server/requirements.txt
# Chimera Phoenix AI Server - Trinity Pipeline Dependencies
# Python 3.8+ required

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI/ML Components
numpy==1.24.3
faiss-cpu==1.7.4  # Use faiss-gpu for GPU support
scikit-learn==1.3.2

# OpenAI Integration (optional - for TCP bridge)
openai==1.3.5  # Optional for Visionary bridge server

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0  # For environment variables
python-multipart==0.0.6  # For file uploads

# Data Processing
pandas==2.1.3  # Optional for corpus analysis
pickle-mixin==1.0.2

# Logging and Monitoring
python-json-logger==2.0.7

# Development Tools (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1  # For testing HTTP endpoints
black==23.11.0  # Code formatting
pylint==3.0.2  # Code linting

# Documentation
mkdocs==1.5.3  # Optional for documentation
mkdocs-material==9.4.8